==> analysis/04-benchmark.Rmd <==
---
title: "Benchmark"
author: "Luke Zappia"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 2
    toc_float: true
    number_sections: true
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
    echo = TRUE,
    fig.width = 10,
    fig.height = 8,
    fig.align = "center"
)
```

```{r libraries}
suppressPackageStartupMessages({
    library(tidyverse)
    library(patchwork)
})
```

```{r source}
source(here::here("analysis", "R", "plotting.R"))
source(here::here("analysis", "R", "summarisation.R"))
```

# Introduction {.unnumbered}

In this document we are going to perform analysis of the benchmark study comparing feature selection methods.

# Overview

First let's get an overview of the metric scores.

```{r load}
metrics <- read_tsv(
    here::here("data", "benchmark.tsv"),
    col_types = cols(
        .default = col_character(),
        Value    = col_double()
    )
) |>
    mutate(
        Type = if_else(Metric == "GraphConnectivity", "IntegrationBio", Type)
    )

metrics_meta <- read_tsv(
    here::here("data", "metrics-metadata.tsv"),
    col_types = cols(
        Metric   = col_character(),
        Name     = col_character(),
        Included = col_logical()
    )
)
include <- metrics_meta$Metric[metrics_meta$Included]
exclude <- metrics_meta$Metric[!metrics_meta$Included]

metric_names <- metrics_meta$Name
names(metric_names) <- metrics_meta$Metric

datasets_meta <- read_tsv(
    here::here("data", "datasets-metadata.tsv"),
    col_types = cols(
        Dataset  = col_character(),
        Name     = col_character(),
    )
)

dataset_names <- datasets_meta$Name
names(dataset_names) <- datasets_meta$Dataset

type_names <- c(
    "IntegrationBatch" = "Integration (Batch)",
    "IntegrationBio" = "Integration (Bio)",
    "Mapping" = "Mapping",
    "Classification" = "Classification",
    "Unseen" = "Unseen populations"
)

metrics
```

## Counts

Check the number of values we have for each metric.
All metrics should have the same number (known missing values are checked below).

```{r counts}
plot_metric_counts(metrics)
```

## Missing values

We expect to have missing values for the `Reconstruction` metric which can't be calculated for the `Symphony` integration and the `CellCycle` metric which can't be calculated for simulated datasets.
Values should be complete for other metrics.

```{r missing}
plot_metric_missing(metrics)
```

# Summarise

Let's compute the summarised metrics using the baseline ranges.

```{r summarise}
baselines <- read_tsv(
    here::here("analysis", "output", "baseline-ranges.tsv"),
    col_types = cols(
        .default = col_character(),
        Lower    = col_double(),
        Upper    = col_double(),
        Range    = col_double()
    )
) |>
    mutate(
        Type = if_else(Metric == "GraphConnectivity", "IntegrationBio", Type)
    )

metrics_summary <- metrics |>
    # Calculate average scores for random methods
    mutate(
        Method = str_remove(Method, "-S[0-9]+$"),
    ) |>
    group_by(Dataset, Method, Integration, Type, Metric) |>
    summarise(Value = mean(Value), .groups = "drop") |>
    # Scale and summarise
    summarise_metrics(baselines)

metrics_summary
```

# Overall scores

## By dataset

```{r overall-dataset}
ggplot(metrics_summary, aes(x = Overall, y = Method, colour = Integration)) +
    geom_point() +
    facet_grid(. ~ Dataset) +
    theme_features()
```

## Summarised

```{r overall-summarised}
methods_meta <- read_tsv(
    here::here("data", "methods-metadata.tsv"),
    col_types = cols(
        Method     = col_character(),
        Name       = col_character(),
        IsBaseline = col_logical()
    )
)
method_names <- methods_meta$Name
names(method_names) <- methods_meta$Method

metrics_summary_plotting <- metrics_summary |>
    filter(Integration == "scVI-1") |>
    pivot_longer(
        cols = Classification:Overall,
        names_to = "Type",
        values_to = "Value"
    ) |>
    mutate(
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        )
    )

metric_means <- metrics_summary_plotting |>
    group_by(Method, Type) |>
    summarise(Value = mean(Value), .groups = "drop")

methods_order <- metric_means |>
    filter(Type == "Overall") |>
    arrange(Value) |>
    pull(Method)

metrics_summary_plotting <- metrics_summary_plotting |>
    arrange(Method) |>
    mutate(
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        )
    )

metric_means <- metric_means |>
    mutate(
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        )
    )

types_palette <- c("#f781bf", RColorBrewer::brewer.pal(5, "Set1"))

ggplot(metrics_summary_plotting, aes(x = Value, y = Method, fill = Type)) +
    annotate(
        "rect",
        xmin = -Inf, xmax = 0, ymin = -Inf, ymax = Inf,
        fill = "red", alpha = 0.1
    ) +
    annotate(
        "rect",
        xmin = 1, xmax = Inf, ymin = -Inf, ymax = Inf,
        fill = "blue", alpha = 0.1
    ) +
    geom_vline(xintercept = 0, colour = "red") +
    geom_vline(xintercept = 1, colour = "blue") +
    geom_linerange(
        data = methods_meta |>
            mutate(
                Method = factor(
                    Method,
                    levels = methods_order,
                    labels = method_names[methods_order]
                )
            ),
        aes(x = NULL, y = Method, fill = NULL, alpha = IsBaseline),
        xmin = -Inf, xmax = Inf,
        linewidth = 3, colour = "grey50",
    ) + 
    ggforce::geom_sina(alpha = 0.5, shape = "circle filled", colour  = "white") +
    geom_point(
        data = metric_means,
        aes(fill = Type),
        shape = "diamond filled", size = 3, colour = "white"
    ) +
    scale_colour_manual(values = types_palette) +
    scale_fill_manual(values = types_palette) +
    scale_alpha_manual(values = c(0, 0.3)) +
    facet_grid(. ~ Type, scales = "free_x", space = "free_x") +
    theme_features()
```

## Ranks

### Bar chart

```{r ranks-barchart}
metric_ranks_means <- metrics_summary |>
    filter(Integration == "scVI-1") |>
    pivot_longer(
        cols = Classification:Overall,
        names_to = "Type",
        values_to = "Value"
    ) |>
    group_by(Dataset, Type) |>
    mutate(Rank = rank(Value)) |>
    group_by(Method, Type) |>
    summarise(
        MeanRank = mean(Rank),
        SDRank = sd(Rank),
        .groups = "drop"
    ) |>
    mutate(
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        ),
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        )
    )

ggplot(metric_ranks_means, aes(x = MeanRank, y = Method, fill = Type)) +
    geom_col() +
    scale_fill_manual(values = types_palette) +
    facet_grid(. ~ Type) +
    theme_features()
```

### Heatmap

```{r ranks-heatmap}
ggplot(metric_ranks_means, aes(x = Type, y = Method, colour = Type)) +
    geom_linerange(
        data = methods_meta |>
            mutate(
                Method = factor(
                    Method,
                    levels = methods_order,
                    labels = method_names[methods_order]
                )
            ) |>
            filter(IsBaseline),
        aes(x = NULL, y = Method, fill = NULL),
        xmin = -Inf, xmax = Inf,
        linewidth = 3, colour = "grey50", alpha = 0.3
    ) +
    geom_point(aes(alpha = MeanRank, size = SDRank), shape = "square") +
    scale_y_discrete(drop = FALSE) +
    scale_colour_manual(values = types_palette) +
    scale_size_continuous(
        trans = "reverse",
        limits = c(max(metric_ranks_means$SDRank), 0),
        range = c(0.1, 2.5)
    ) +
    theme_features() +
    theme(
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)
    )
```

# Relationships

## Batch vs Bio

```{r batch-bio}
meta_types <- metrics_summary |>
    filter(Integration == "scVI-1") |>
    mutate(
        Reference = 0.5 * IntegrationBatch + 0.5 * IntegrationBio,
        Query = Mapping / 3 + Classification / 3 + Unseen / 3,
        Batch = IntegrationBatch / 2 + Mapping / 2,
        Bio = IntegrationBio / 3 + Classification / 3 + Unseen / 3
    ) |>
    group_by(Dataset) |>
    mutate(RankOverall = rank(Overall)) |>
    ungroup()

ggplot(meta_types, aes(x = Batch, y = Bio)) +
    geom_point(aes(colour = RankOverall)) +
    geom_smooth(aes(fill = Dataset), method = "lm", formula = "y ~ x") + 
    scale_colour_viridis_c() +
    facet_wrap(~ Dataset) +
    theme_features()
```

## Query vs References

```{r query-reference}
ggplot(meta_types, aes(x = Reference, y = Query)) +
    geom_point(aes(colour = RankOverall)) +
    geom_smooth(aes(fill = Dataset), method = "lm", formula = "y ~ x") +
    scale_colour_viridis_c() +
    facet_wrap(~ Dataset) +
    theme_features()
```

# Batch aware

## Line plot

```{r batch-lineplot}
batch_methods_means <- metric_means |>
    filter(str_detect(Method, "scanpy")) |>
    mutate(Batch = str_detect(Method, "Batch=True")) |>
    mutate(
        Method = str_remove(Method, "\\s\\(.*\\)")
    )

ggplot(batch_methods_means, aes(x = Batch, y = Value, colour = Method)) +
    geom_line(aes(group = Method)) +
    geom_point(size = 2) +
    facet_wrap(~ Type, scales = "free_y", nrow = 1) +
    theme_features()
```

## Heatmap

```{r batch-heatmap}
batch_methods_diffs <- metrics_summary_plotting |>
    filter(str_detect(Method, "scanpy")) |>
    mutate(Batch = str_detect(Method, "Batch=True")) |>
    mutate(
        Method = str_remove(Method, "\\s\\(.*\\)")
    ) |>
    pivot_wider(
        names_from = "Batch", names_prefix = "Batch", values_from = "Value",
    ) |>
    mutate(
        Difference = BatchTRUE - BatchFALSE
    )

methods_clust <- batch_methods_diffs |>
    select(Dataset, Type, Method, Difference) |>
    pivot_wider(
        id_cols = Method,
        names_from = c(Dataset, Type),
        values_from = Difference
    ) |>
    column_to_rownames("Method") |>
    as.matrix() |>
    dist() |>
    hclust()
methods_clust_order <- methods_clust$labels[methods_clust$order]

datasets_clust <- batch_methods_diffs |>
    select(Dataset, Type, Method, Difference) |>
    pivot_wider(
        id_cols = Dataset,
        names_from = c(Method, Type),
        values_from = Difference
    ) |>
    column_to_rownames("Dataset") |>
    as.matrix() |>
    dist() |>
    hclust()
datasets_clust_order <- datasets_clust$labels[datasets_clust$order]

batch_methods_diffs |>
    mutate(
        Method = factor(
            Method,
            levels = methods_clust_order
        ),
        Dataset = factor(
            Dataset,
            levels = datasets_clust_order
        )
    ) |>
    ggplot(aes(x = Dataset, y = Method, fill = Difference)) +
        geom_tile() +
        colorspace::scale_fill_continuous_diverging(
            palette = "PurpleGreen"
        ) +
        facet_wrap(~ Type) +
        theme_features() +
        theme(
            axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)
        )
```

# Selected features

## Overlap

```{r selected-overlap}
feature_files <- fs::dir_ls(
    here::here("data", "selected-features"),
    glob = "*.tsv"
)

selected_list <- map(feature_files, function(.file) {
    read_tsv(
        .file,
        col_types = cols(
            .default = col_logical(),
            Feature = col_character()
        )
    ) |>
        column_to_rownames("Feature") |>
        as.matrix()
})
names(selected_list) <- fs::path_ext_remove(fs::path_file(names(selected_list)))

overlaps <- map_dfr(names(selected_list), function(.dataset) {
    selected <- selected_list[[.dataset]]
    expand_grid(
        method1 = colnames(selected),
        method2 = colnames(selected)
    ) |>
        filter(
            str_detect(method1, "random", negate = TRUE),
            str_detect(method2, "random", negate = TRUE)
        ) |>
        pmap_dfr(function(method1, method2) {
            selected1 <- selected[, method1]
            selected2 <- selected[, method2]
            
            both <- sum(selected1 & selected2)
            either <- sum(selected1 | selected2)
            subset <- sum(selected1[selected2])
            
            tibble(
                Dataset = .dataset,
                Method1 = method1,
                Method2 = method2,
                Both = both,
                Either = either,
                Jaccard = both / either,
                Subset = subset,
                Prop = subset / sum(selected1)
            )
        })
})

overlap_means <- overlaps |>
    group_by(Method1, Method2) |>
    summarise(
        JaccardSD = sd(Jaccard),
        Jaccard = mean(Jaccard),
        PropSD = sd(Prop),
        Prop = mean(Prop),
        .groups = "drop"
    )

overlaps_clust <- overlap_means |>
    select(Method1, Method2, Jaccard) |>
    pivot_wider(names_from = Method2, values_from = Jaccard) |>
    column_to_rownames("Method1") |>
    as.matrix() |>
    dist() |>
    hclust()
overlaps_clust_order <- overlaps_clust$labels[overlaps_clust$order]

overlap_means |>
    # mutate(
    #     Method1 = factor(
    #         str_replace_all(Method1, "_", "-"),
    #         levels = methods_order
    #     ),
    #     Method2 = factor(
    #         str_replace_all(Method2, "_", "-"),
    #         levels = methods_order
    #     )
    # ) |>
    mutate(
        Method1 = factor(
            Method1,
            levels = overlaps_clust_order
        ),
        Method2 = factor(
            Method2,
            levels = overlaps_clust_order
        )
    ) |>
    ggplot(aes(x = Method1, y = Method2, fill = Jaccard)) +
        geom_tile() +
        scale_fill_viridis_c() +
        theme_features() +
        theme(
            axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
            axis.text.y = element_text(angle = 0, hjust = 1, vjust = 0.5)
        )
```

## Number

```{r selected-number}
as.data.frame(colSums(selected_list[[1]]))

num_selected <- map_dfr(names(selected_list), function(.dataset) {
    selected <- selected_list[[.dataset]]
    col_sums <- colSums(selected)
    tibble(
        Dataset = .dataset,
        Method = names(col_sums),
        Selected = col_sums
    )
})

num_selected_means <- num_selected |>
    group_by(Method) |>
    summarise(
        MeanSelected = mean(Selected),
        SDSelected = sd(Selected),
        .groups = "drop"
    )

ggplot(num_selected, aes(x = Selected, y = Method)) +
    geom_vline(xintercept = 2000, colour = "red") +
    geom_point(
        aes(colour = Dataset),
        position = position_jitter(width = 0, height = 0.2, seed = 1),
        alpha = 0.5
    ) +
    geom_point(
        data = num_selected_means,
        aes(x = MeanSelected),
        shape = "|", size = 4
    ) +
    scale_x_log10() +
    theme_features()
```

## Number of methods

### Count

Number of methods that select each feature

```{r selected-num-methods}
num_methods <- map_dfr(names(selected_list), function(.dataset) {
    selected <- selected_list[[.dataset]]
    row_sums <- rowSums(selected)
    count_table <- table(row_sums)
    tibble(
        Dataset = .dataset,
        NumMethods = names(count_table),
        NumFeatures = as.numeric(count_table)
    )
}) |>
    mutate(NumMethods = as.numeric(NumMethods))

ggplot(num_methods, aes(x = NumMethods, y = NumFeatures)) +
    geom_vline(xintercept = 10, colour = "red") +
    geom_vline(xintercept = 25, colour = "blue") +
    geom_point() +
    facet_wrap(~ Dataset) +
    scale_y_log10() +
    theme_features()
```

### Cumulative

Number of features that are selected by at least a given number of methods

```{r selected-cumulative}
cum_num_methods <- num_methods |>
    arrange(Dataset, desc(NumMethods)) |>
    group_by(Dataset) |>
    mutate(
        CumFeatures = cumsum(NumFeatures)
    )

ggplot(cum_num_methods, aes(x = NumMethods, y = CumFeatures)) +
    geom_vline(xintercept = 10, colour = "red") +
    geom_vline(xintercept = 25, colour = "blue") +
    geom_point() +
    facet_wrap(~ Dataset) +
    scale_y_log10() +
    theme_features()
```

# Integration methods

## Line plot

```{r integration-lineplot}
integration_ranks <- metrics_summary |>
    pivot_longer(
        cols = Classification:Overall,
        names_to = "Type",
        values_to = "Value"
    ) |>
    group_by(Dataset, Integration, Type) |>
    mutate(Rank = rank(Value))

integration_ranks_means <- integration_ranks |>
    group_by(Integration, Method, Type) |>
    summarise(
        MeanRank = mean(Rank),
        SDRank = sd(Rank),
        .groups = "drop"
    )

ggplot(integration_ranks, aes(x = Integration, y = Rank)) +
    geom_line(aes(group = Method)) +
    geom_point() +
    facet_grid(Dataset ~ Type) +
    theme_features()
```

## Heatmap

```{r integration-heatmap}
integration_ranks |>
    mutate(
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        )
    ) |>
    ggplot(aes(x = Method, y = Integration, fill = Rank)) +
        geom_tile() + 
        scale_fill_viridis_c() +
        facet_grid(Dataset ~ Type) +
        theme_features() +
        theme(
            axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)
        )
```

## Scatter plot {.tabset}

### scANVI {.unnumbered}

```{r integration-scatter-scANVI}
integration_summary <- metrics_summary |>
    pivot_longer(
        cols = Classification:Overall,
        names_to = "Type",
        values_to = "Value"
    ) |>
    mutate(
        Integration = str_remove(Integration, "-[0-9]+$")
    ) |>
    pivot_wider(
        names_from = "Integration",
        values_from = "Value"
    )

ggplot(integration_summary, aes(x = scVI, y = scANVI)) +
    geom_abline(intercept = 0, slope = 1, colour = "red") +
    geom_point(aes(colour = Dataset)) +
    facet_grid(Dataset ~ Type, scales = "free") +
    theme_features()
```

### Symphony {.unnumbered}

```{r integration-scatter-Symphony}
ggplot(integration_summary, aes(x = scVI, y = Symphony)) +
    geom_abline(intercept = 0, slope = 1, colour = "red") +
    geom_point(aes(colour = Dataset)) +
    facet_grid(Dataset ~ Type, scales = "free") +
    theme_features()
```

## Difference scatter {.tabset}

### scANVI {.unnumbered}

```{r integration-scatter-diff-scANVI}
integration_diffs <- integration_summary |>
    mutate(
        scANVIDiff = scANVI - scVI,
        SymphonyDiff = Symphony - scVI
    )

ggplot(integration_diffs, aes(x = scVI, y = scANVIDiff)) +
    geom_hline(yintercept = 0, colour = "red") +
    geom_point(aes(colour = Dataset)) +
    facet_grid(Dataset ~ Type, scales = "free") +
    theme_features()

```

### Symphony {.unnumbered}

```{r integration-scatter-diff-Symphony}
ggplot(integration_diffs, aes(x = scVI, y = SymphonyDiff)) +
    geom_hline(yintercept = 0, colour = "red") +
    geom_point(aes(colour = Dataset)) +
    facet_grid(Dataset ~ Type, scales = "free") +
    theme_features()
```

## Difference heatmap {.tabset}

### scANVI {.unnumbered}

```{r integration-heatmap-diff-scANVI}
integration_diff_means <- integration_diffs |>
    group_by(Method, Type) |>
    summarise(
        scANVIMean = mean(scANVIDiff),
        scANVISD = sd(scANVIDiff),
        SymphonyMean = mean(SymphonyDiff),
        SymphonySD = sd(SymphonyDiff),
        .groups = "drop"
    )

integration_diff_means |>
    mutate(
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        )
    ) |>
    ggplot(aes(x = Type, y = Method, fill = scANVIMean)) +
        geom_tile() +
        colorspace::scale_fill_continuous_diverging(
            palette = "PurpleGreen"
        ) +
        theme_features()
```

### Symphony {.unnumbered}

```{r integration-heatmap-diff-Symphony}
integration_diff_means |>
    mutate(
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        )
    ) |>
    ggplot(aes(x = Type, y = Method, fill = SymphonyMean)) +
        geom_tile() +
        colorspace::scale_fill_continuous_diverging(
            palette = "PurpleGreen"
        ) +
        theme_features()
```

## Rank differences scatter {.tabset}

### scANVI {.unnumbered}

```{r integration-scatter-rank-diff-scANVI}
integration_rank_diffs <- integration_ranks |>
    mutate(
        Integration = str_remove(Integration, "-[0-9]+$")
    ) |>
    select(-Value) |>
    pivot_wider(
        names_from = "Integration",
        values_from = "Rank"
    ) |>
    mutate(
        scANVIDiff = scANVI - scVI,
        SymphonyDiff = Symphony - scVI
    )

ggplot(integration_rank_diffs, aes(x = scVI, y = scANVIDiff)) +
    geom_hline(yintercept = 0, colour = "red") +
    geom_point(aes(colour = Dataset)) +
    facet_grid(Dataset ~ Type, scales = "free") +
    theme_features()
```

### Symphony {.unnumbered}

```{r integration-scatter-rank-diff-Symphony}
ggplot(integration_rank_diffs, aes(x = scVI, y = SymphonyDiff)) +
    geom_hline(yintercept = 0, colour = "red") +
    geom_point(aes(colour = Dataset)) +
    facet_grid(Dataset ~ Type, scales = "free") +
    theme_features()
```

## Rank differences heatmap {.tabset}

### scANVI {.unnumbered}

```{r integration-heatmap-rank-diff-scANVI}
integration_ranks_diff_means <- integration_rank_diffs |>
    group_by(Method, Type) |>
    summarise(
        scANVIMean = mean(scANVIDiff),
        scANVISD = sd(scANVIDiff),
        SymphonyMean = mean(SymphonyDiff),
        SymphonySD = sd(SymphonyDiff),
        .groups = "drop"
    )

integration_ranks_diff_means |>
    mutate(
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        )
    ) |>
    ggplot(aes(x = Type, y = Method, fill = scANVIMean)) +
        geom_tile() +
        colorspace::scale_fill_continuous_diverging(
            palette = "PurpleGreen"
        ) +
        theme_features()
```

### Symphony {.unnumbered}

```{r integration-heatmap-rank-diff-Symphony}
integration_ranks_diff_means |>
    mutate(
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        )
    ) |>
    ggplot(aes(x = Type, y = Method, fill = SymphonyMean)) +
        geom_tile() +
        colorspace::scale_fill_continuous_diverging(
            palette = "PurpleGreen"
        ) +
        theme_features()
```

# Figures

## Benchmark

```{r figure-benchmark}
# Values plot | rankings heatmap
overall_values_figure <- ggplot(
    mutate(
        metrics_summary_plotting,
        Type = fct_recode(
            Type,
            "Int. (Batch)" = "Integration (Batch)",
            "Int. (Bio)" = "Integration (Bio)",
            "Unseen" = "Unseen populations"
        )
    ),
    aes(x = Value, y = Method, fill = Type)
) +
    annotate(
        "rect",
        xmin = -Inf, xmax = 0, ymin = -Inf, ymax = Inf,
        fill = "red", alpha = 0.1
    ) +
    annotate(
        "rect",
        xmin = 1, xmax = Inf, ymin = -Inf, ymax = Inf,
        fill = "blue", alpha = 0.1
    ) +
    geom_vline(xintercept = 0, colour = "red") +
    geom_vline(xintercept = 1, colour = "blue") +
    geom_linerange(
        data = methods_meta |>
            mutate(
                Method = factor(
                    Method,
                    levels = methods_order,
                    labels = method_names[methods_order]
                )
            ),
        aes(x = NULL, y = Method, fill = NULL, alpha = IsBaseline),
        xmin = -Inf, xmax = Inf,
        linewidth = 3, colour = "grey50",
    ) + 
    ggforce::geom_sina(
        size = 0.8, alpha = 0.5, shape = "circle filled", colour  = "white"
    ) +
    geom_point(
        data = mutate(
            metric_means,
            Type = fct_recode(
                Type,
                "Int. (Batch)" = "Integration (Batch)",
                "Int. (Bio)" = "Integration (Bio)",
                "Unseen" = "Unseen populations"
            )
        ),
        aes(fill = Type),
        shape = "diamond filled", size = 1.2, colour = "white"
    ) +
    scale_colour_manual(values = types_palette) +
    scale_fill_manual(values = types_palette) +
    scale_alpha_manual(values = c(0, 0.3)) +
    facet_grid(. ~ Type, scales = "free_x") +
    labs(
        alpha = "Baseline method"
    ) +
    theme_features_pub() +
    theme(
        legend.position = "bottom",
        axis.title = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)
    )

overall_ranks_figure <- ggplot(
    metric_ranks_means,
    aes(x = Type, y = Method, colour = Type)
) +
    geom_linerange(
        data = methods_meta |>
            mutate(
                Method = factor(
                    Method,
                    levels = methods_order,
                    labels = method_names[methods_order]
                )
            ) |>
            filter(IsBaseline),
        aes(x = NULL, y = Method, fill = NULL),
        xmin = -Inf, xmax = Inf,
        linewidth = 3, colour = "grey50", alpha = 0.3
    ) +
    geom_point(
        aes(alpha = length(unique(Method)) -  MeanRank, size = SDRank),
        shape = "square"
    ) +
    scale_y_discrete(drop = FALSE) +
    scale_colour_manual(values = types_palette, guide = "none") +
    scale_size_continuous(
        trans = "reverse",
        limits = c(max(metric_ranks_means$SDRank), 0),
        range = c(0.1, 1.8)
    ) +
    scale_alpha_continuous(
        limits = c(length(unique(metric_ranks_means$Method)), 1),
        range = c(0.1, 1.0),
        trans = "reverse"
    ) +
    labs(
        title = "Mean ranks",
        alpha = "Mean rank",
        size = "Rank SD"
    ) +
    theme_features_pub() +
    theme(
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        axis.title = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.y = element_blank()
    )

overall_panel <- wrap_plots(
    overall_values_figure,
    overall_ranks_figure,
    nrow = 1,
    widths = c(5, 0.8)
)

# Features overlap | Number of methods | Number of features

overlaps_clust_order_formatted <- overlaps_clust_order |>
    str_replace_all("_", "-") |>
    str_replace("seurat-v3", "seurat_v3") |>
    str_replace("cell-ranger", "cell_ranger")

overlaps_plotting <- overlap_means |>
    # mutate(
    #     Method1 = factor(
    #         str_replace_all(Method1, "_", "-"),
    #         levels = str_replace_all(methods_order, "_", "-")
    #     ),
    #     Method2 = factor(
    #         str_replace_all(Method2, "_", "-"),
    #         levels = str_replace_all(methods_order, "_", "-")
    #     )
    # ) |>
    mutate(
        Method1 = factor(
            Method1,
            levels = overlaps_clust_order,
            labels = str_remove(
                str_remove(
                    method_names[overlaps_clust_order_formatted],
                    "N=2000,?\\s?"
                ),
                "\\s\\(\\)"
            )
        ),
        Method2 = factor(
            Method2,
            levels = overlaps_clust_order,
            labels = str_remove(
                str_remove(
                    method_names[overlaps_clust_order_formatted],
                    "N=2000,?\\s?"
                ),
                "\\s\\(\\)"
            )
        )
    )

mean_overlaps_plot <- ggplot(
    overlaps_plotting,
    aes(
        x = Method1, y = Method2,
        fill = Jaccard, colour = Jaccard > 0.5, size = JaccardSD
    )
) +
    geom_point(shape = "square filled", stroke = 0.3) +
    scale_fill_viridis_c(option = "plasma", limits = c(0, 1)) +
    scale_colour_manual(values = c("black", "white")) +
    scale_size_continuous(
        trans = "reverse",
        range = c(0.1, 1.8), breaks = scales::breaks_extended(n = 3)
    ) +
    coord_equal() +
    labs(
        title = "Average overlap",
        fill = "Mean\nJaccard\nIndex",
        colour = "JI > 0.5",
        size = "Standard\ndeviation"
    ) +
    guides(
        fill = guide_colourbar(order = 1),
        colour = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            order = 2
        ),
        size = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            order = 3
        )
    ) +
    theme_features_pub() +
    theme(
        legend.position = "bottom",
        axis.text.x = element_blank(),
        axis.title = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill = "black"),
        panel.grid = element_blank()
    )

cum_num_methods_plotting <- cum_num_methods |>
    filter(NumMethods %in% c(5, 10, 15, 20, 25)) |>
    mutate(
        Dataset = factor(
            Dataset,
            levels = names(dataset_names),
            labels = dataset_names
        )
    )

num_methods_plot <- ggplot(
    cum_num_methods_plotting,
    aes(y = Dataset, x = CumFeatures, fill = factor(NumMethods))
) +
    geom_point(size = 1, shape = "circle filled") +
    scale_fill_viridis_d(option = "plasma") +
    scale_x_log10(breaks = c(10, 100, 500, 1000, 5000, 10000)) +
    labs(
        x = "Number of selected features (log scale)",
        fill = "Number of\nmethods"
    ) +
    theme_features_pub() +
    theme(
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        axis.title.y = element_blank(),
        panel.grid.minor = element_blank()
    )

num_selected_means_plotting <- num_selected_means |>
    filter(SDSelected > 0) |>
    mutate(
        Method = str_replace_all(Method, "_", "-"),
        MethodLabel = factor(
            Method,
            levels = names(method_names),
            labels = str_replace(method_names,  " ", "\n")
        )
    ) |>
    arrange(MeanSelected) |>
    mutate(MethodLabel = fct_reorder(MethodLabel, MeanSelected))

num_selected_plotting <- num_selected |>
    mutate(Method = str_replace_all(Method, "_", "-")) |>
    filter(Method %in% num_selected_means_plotting$Method) |>
    mutate(
        MethodLabel = factor(
            Method,
            levels = num_selected_means_plotting$Method,
            labels = num_selected_means_plotting$MethodLabel
        ),
        Dataset = factor(
            Dataset,
            levels = names(dataset_names),
            labels = dataset_names
        )
    )

num_selected_plot <- ggplot(
    num_selected_plotting,
    aes(x = Selected, y = MethodLabel)
) +
    geom_vline(xintercept = 2000, colour = "red") +
    geom_point(
        aes(fill = Dataset),
        position = position_jitter(width = 0, height = 0.2, seed = 1),
        size = 1, shape = "circle filled", alpha = 0.8, colour = "grey20"
    ) +
    geom_point(
        data = num_selected_means_plotting,
        aes(x = MeanSelected),
        shape = "|", size = 4, colour = "blue"
    ) +
    scale_x_log10(
        breaks = c(100, 200, 500, 1000, 2000, 5000, 10000, 20000)
    ) +
    scale_fill_brewer(palette = "Set3") +
    labs(
        x = "Number of selected\nfeatures (log scale)"
    ) +
    theme_features_pub() +
    theme(
        axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        panel.grid.minor.x = element_blank()
    )

overlaps_panel <- wrap_plots(
    mean_overlaps_plot,
    # dataset_overlaps_plot,
    wrap_plots(num_methods_plot, num_selected_plot, guides = "collect"),
    nrow = 1,
    widths = c(1.0, 1.1)
)

# Batch aware difference heatmap
batch_aware_plot <- batch_methods_diffs |>
    mutate(
        Method = factor(
            Method,
            levels = methods_clust_order
        ),
        Dataset = factor(
            Dataset,
            levels = datasets_clust_order,
            labels = dataset_names[datasets_clust_order]
        )
    ) |>
    ggplot(aes(x = Dataset, y = Method, fill = Difference)) +
        geom_tile() +
        colorspace::scale_fill_continuous_diverging(
            palette = "PurpleGreen"
        ) +
        scale_y_discrete(expand = c(0, 0)) +
        facet_wrap(~ Type, nrow = 1) +
        labs(
            fill = "Batch aware -\nstandard"
        ) +
        theme_features_pub() +
        theme(
            axis.title = element_blank(),
            axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5)
        )

benchmark_figure <- wrap_plots(
    overall_panel,
    overlaps_panel,
    batch_aware_plot,
    nrow = 3,
    heights = c(3.5, 2.2, 0.6)
)

benchmark_figure
```

## Integration

```{r figure-integration}
theme_features_integration <- theme_features_pub() +
    theme(
        legend.position = "bottom",
        legend.title.position = "top",
        panel.grid = element_blank(),
        panel.spacing.x = unit(0, "cm"),
        axis.title = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        axis.ticks.y = element_blank(),
        legend.key.width = unit(0.35, "cm"),
        legend.key.spacing.x = unit(0.0, "cm"),
        legend.box = "vertical"
    )

integration_summary_plotting <- integration_summary |>
    pivot_longer(
        Symphony:scVI,
        names_to = "Integration",
        values_to = "Value"
    ) |>
    group_by(Method, Type, Integration) |>
    summarise(
        Mean = mean(Value),
        SD = sd(Value),
        .groups = "drop"
    ) |>
    mutate(
        Integration = factor(
            Integration,
            levels = c("scVI", "scANVI", "Symphony")
        ),
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        ),
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        )
    )

integration_scores_plot <- ggplot(
    integration_summary_plotting,
    aes(x = Type, y = Method, colour = Mean, size = SD)
) +
    geom_point(shape = "square") +
    scale_colour_viridis_c(
        option = "magma",
        limits = c(0, max(integration_summary_plotting$Mean))
    ) +
    scale_size_continuous(
        trans = "reverse", range = c(0.1, 2),
        breaks = scales::breaks_extended(n = 5)
    ) +
    labs(
        title = "Mean scores",
        colour = "Mean score",
        size = "Standard deviation",
    ) +
    guides(
        colour = guide_colourbar(order = 1),
        size = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            order = 2
        )
    ) +
    facet_wrap(~ Integration, nrow = 1) +
    theme_features_integration +
    theme(
        axis.ticks.y = element_line(size = 0.25)
    )

# Integration difference heatmap
integration_diff_means_plotting <- integration_diff_means |>
    pivot_longer(scANVIMean:SymphonySD) |>
    mutate(
        Integration = if_else(
            str_detect(name, "scANVI"),
            "scANVI",
            "Symphony"
        ),
        Summary = if_else(
            str_detect(name, "Mean"),
            "Mean",
            "SD"
        )
    ) |>
    select(-name) |> 
    pivot_wider(
        names_from = Summary,
        values_from = value
    ) |>
    mutate(
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        ),
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        )
    )

integration_scores_diffs_plot <- ggplot(
    integration_diff_means_plotting,
    aes(x = Type, y = Method, colour = Mean, size = SD)
) +
    geom_point(shape = "square") +
    colorspace::scale_colour_continuous_diverging(
        palette = "PurpleGreen",
    ) +
    scale_size_continuous(
        trans = "reverse", range = c(0.1, 2),
        breaks = scales::breaks_extended(n = 5)
    ) +
    facet_wrap(~ Integration, nrow = 1) +
    labs(
        title = "Difference in\nmean scores",
        colour = "Difference to scVI",
        size = "Standard deviation"
    ) +
    guides(
        colour = guide_colourbar(order = 1),
        size = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            order = 2
        )
    ) +
    theme_features_integration +
    theme(
        axis.text.y = element_blank(),
    )

integration_ranks_means_plotting <- integration_ranks_means |>
    mutate(Integration = str_remove(Integration, "-[0-9]+$")) |>
    mutate(
        Integration = factor(
            Integration,
            levels = c("scVI", "scANVI", "Symphony")
        ),
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        ),
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        )
    )

integration_ranks_plot <- ggplot(
    integration_ranks_means_plotting,
    aes(x = Type, y = Method, colour = Type)
) +
    geom_point(
        aes(alpha = length(unique(Method)) -  MeanRank, size = SDRank),
        shape = "square"
    ) +
    scale_y_discrete(drop = FALSE) +
    scale_colour_manual(values = types_palette, guide = "none") +
    scale_size_continuous(
        trans = "reverse",
        limits = c(max(integration_ranks_means$SDRank), 0),
        range = c(0.1, 2), breaks = scales::breaks_extended(n = 6)
    ) +
    scale_alpha_continuous(
        limits = c(length(unique(integration_ranks_means$Method)), 1),
        range = c(0.1, 1.0),
        trans = "reverse", breaks = scales::breaks_extended(n = 6)
    ) +
    facet_wrap(~ Integration, nrow = 1) +
    labs(
        title = "Mean ranks",
        alpha = "Mean rank",
        size = "Standard deviation"
    ) +
    guides(
        alpha = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            nrow = 1,
            order = 1
        ),
        size = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            order = 2
        )
    ) +
    theme_features_integration +
    theme(
        axis.text.y = element_blank()
    )

integration_ranks_diff_means_plotting <- integration_ranks_diff_means |>
    pivot_longer(scANVIMean:SymphonySD) |>
    mutate(
        Integration = if_else(
            str_detect(name, "scANVI"),
            "scANVI",
            "Symphony"
        ),
        Summary = if_else(
            str_detect(name, "Mean"),
            "Mean",
            "SD"
        )
    ) |>
    select(-name) |> 
    pivot_wider(
        names_from = Summary,
        values_from = value
    ) |>
    mutate(
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        ),
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        )
    )

integration_diffs_ranks_plot <- ggplot(
    integration_ranks_diff_means_plotting,
    aes(x = Type, y = Method, colour = Mean, size = SD)
) +
    geom_point(shape = "square") +
    colorspace::scale_colour_continuous_diverging(
        palette = "Tropic",
        rev = TRUE
    ) +
    scale_size_continuous(
        trans = "reverse", range = c(0.1, 2),
        breaks = scales::breaks_extended(n = 5)
    ) +
    facet_wrap(~ Integration, nrow = 1) +
    labs(
        title = "Difference in\nmean ranks",
        colour = "Difference to scVI",
        size = "Standard deviation"
    ) +
    guides(
        colour = guide_colourbar(order = 1),
        size = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            nrow = 1,
            order = 2
        )
    ) +
    theme_features_integration +
    theme(
        axis.text.y = element_blank(),
    )

integration_figure <- wrap_plots(
    integration_scores_plot,
    integration_scores_diffs_plot,
    integration_ranks_plot,
    integration_diffs_ranks_plot,
    nrow = 1,
    widths = c(3, 2, 3, 2)
)

integration_figure
```

# Extended figures

## Dataset scores and ranks

```{r extended-scores-ranks, fig.width = 8, out.width = "100%"}
category_scores_plot <- ggplot(
    metrics_summary_plotting |>
        mutate(
            Method = fct_relabel(
                Method,
                \(.x) {
                    str_remove(.x, " \\(N=2000\\)") |>
                        str_remove("N=2000, ") |>
                        str_wrap(width = 15)
                }
            ),
            Type = fct_recode(
                Type,
                "Int. (Batch)" = "Integration (Batch)",
                "Int. (Bio)" = "Integration (Bio)",
                "Unseen" = "Unseen populations"
            )
        ),
        Dataset = factor(
            Dataset,
            levels = names(dataset_names),
            labels = dataset_names
        ),
    aes(x = Dataset, y = Method)
) +
    geom_linerange(
        data = methods_meta |>
            mutate(
                Method = factor(
                    Method,
                    levels = methods_order,
                    labels = method_names[methods_order]
                ) |>
                    fct_relabel(
                        \(.x) {
                            str_remove(.x, " \\(N=2000\\)") |>
                                str_remove("N=2000, ") |>
                                str_wrap(width = 15)
                        }
                    )
            ) |>
            filter(IsBaseline),
        aes(x = NULL, y = Method, fill = NULL),
        xmin = -Inf, xmax = Inf,
        linewidth = 3, colour = "grey50", alpha = 0.3
    ) +
    geom_point(aes(colour = Value), shape = "square", size = 1) +
    scale_y_discrete(drop = FALSE) +
    colorspace::scale_colour_continuous_divergingx(
        palette = "Zissou 1",
        limits = c(
            -max(abs(metrics_summary_plotting$Value)),
            max(abs(metrics_summary_plotting$Value))
        )
    ) +
    facet_grid(. ~ Type) +
    labs(
        title = "Metric category scores",
        colour = "Metric category score"
    ) +
    theme_features_pub() +
    theme(
        legend.position = "bottom",
        axis.text.x = element_text(size = 4, angle = 90, hjust = 1, vjust = 0.5),
        axis.text.y = element_text(size = 4),
        axis.title = element_blank()
    )

metric_ranks <- metrics_summary |>
    filter(Integration == "scVI-1") |>
    pivot_longer(
        cols = Classification:Overall,
        names_to = "Type",
        values_to = "Value"
    ) |>
    group_by(Dataset, Type) |>
    mutate(Rank = rank(Value)) |>
    mutate(
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        ) |>
            fct_recode(
                "Int. (Batch)" = "Integration (Batch)",
                "Int. (Bio)" = "Integration (Bio)",
                "Unseen" = "Unseen populations"
            ),
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        ) |>
            fct_relabel(
                \(.x) {
                    str_remove(.x, " \\(N=2000\\)") |>
                        str_remove("N=2000, ") |>
                        str_wrap(width = 15)
                }
            ),
        Dataset = factor(
            Dataset,
            levels = names(dataset_names),
            labels = dataset_names
        )
    )

category_ranks_plot <- ggplot(
    metric_ranks,
    aes(x = Dataset, y = Method, colour = Type)
) +
    geom_linerange(
        data = methods_meta |>
            mutate(
                Method = factor(
                    Method,
                    levels = methods_order,
                    labels = method_names[methods_order]
                ) |>
                    fct_relabel(
                        \(.x) {
                            str_remove(.x, " \\(N=2000\\)") |>
                                str_remove("N=2000, ") |>
                                str_wrap(width = 15)
                        }
                    ),
            ) |>
            filter(IsBaseline),
        aes(x = NULL, y = Method, fill = NULL),
        xmin = -Inf, xmax = Inf,
        linewidth = 3, colour = "grey50", alpha = 0.3
    ) +
    geom_point(
        aes(alpha = length(unique(Method)) -  Rank),
        shape = "square", size = 1
    ) +
    scale_y_discrete(drop = FALSE) +
    scale_colour_manual(values = types_palette, guide = "none") +
    scale_alpha_continuous(
        limits = c(length(unique(metric_ranks$Method)), 1),
        range = c(0.1, 1.0),
        trans = "reverse",
        breaks = scales::extended_breaks(n = 10)
    ) +
    facet_grid(. ~ Type) +
    labs(
        title = "Metric category ranks",
        alpha = "Mean category rank",
    ) +
    guides(
        alpha = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            nrow = 1,
        )
    ) +
    theme_features_pub() +
    theme(
        legend.position = "bottom",
        axis.text.x = element_text(size = 4, angle = 90, hjust = 1, vjust = 0.5),
        axis.text.y = element_blank(),
        axis.title = element_blank(),
        axis.ticks.y = element_blank()
    )

category_scores_ranks_plot <- wrap_plots(
    category_scores_plot,
    category_ranks_plot,
    nrow = 1,
    widths = c(1, 1)
)

category_scores_ranks_plot
```

## Overlaps means and SDs

```{r extended-overlaps-means-sds, fig.width = 8, out.width = "100%"}
overlaps_means_plot <- ggplot(
    overlaps_plotting,
    aes(x = Method1, y = Method2, fill = Jaccard)
) +
    geom_tile() +
    scale_fill_viridis_c(option = "plasma", limits = c(0, 1)) +
    scale_x_discrete(position = "top") +
    coord_equal() +
    labs(fill = "Mean\nJaccard\nIndex",) +
    guides(
        fill = guide_colourbar(order = 1),
        colour = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            order = 2
        )
    ) +
    theme_features_pub() +
    theme(
        legend.position = "bottom",
        axis.text.x = element_text(angle = 90, hjust = 0, vjust = 0.5),
        axis.title = element_blank(),
        panel.background = element_rect(fill = "black"),
        panel.grid = element_blank()
    )

overlaps_sds_plot <- ggplot(
    overlaps_plotting,
    aes(x = Method1, y = Method2, fill = JaccardSD)
) +
    geom_tile() +
    scale_fill_viridis_c(option = "cividis") +
    scale_x_discrete(position = "top") +
    coord_equal() +
    labs(fill = "Jaccard\nIndex\nSD",) +
    guides(
        fill = guide_colourbar(order = 1),
        colour = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            order = 2
        )
    ) +
    theme_features_pub() +
    theme(
        legend.position = "bottom",
        axis.text.y = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 0, vjust = 0.5),
        axis.title = element_blank(),
        panel.background = element_rect(fill = "black"),
        panel.grid = element_blank()
    )

overlaps_means_sd_plot <- wrap_plots(
    overlaps_means_plot, overlaps_sds_plot,
    nrow = 1,
    widths = c(1, 1)
)

overlaps_means_sd_plot
```

## Dataset overlaps

```{r extended-dataset-overlaps, fig.width = 8, out.width = "100%"}
dataset_overlaps_plot <- overlaps |>
    mutate(
        Method1 = factor(
            Method1,
            levels = overlaps_clust_order
        ),
        Method2 = factor(
            Method2,
            levels = overlaps_clust_order
        ),
        Dataset = factor(
            Dataset,
            levels = names(dataset_names),
            labels = dataset_names
        )
    ) |>
ggplot(
    aes(x = Method1, y = Method2, fill = Jaccard)
) +
    geom_tile() +
    facet_wrap(~ Dataset, nrow = 2) +
    scale_fill_viridis_c(option = "plasma", limits = c(0, 1)) +
    scale_colour_manual(values = c("black", "white")) +
    coord_equal() +
    theme_features_pub() +
    theme(
        legend.position = "none",
        axis.text = element_blank(),
        axis.title = element_blank(),
        panel.background = element_rect(fill = "black"),
        panel.grid = element_blank()
    )

dataset_overlaps_plot
```

## Integration means

```{r extended-integration-means}
theme_features_integration_extended <- theme_features_integration +
    theme(
        axis.text.x = element_text(size = 5, angle = 90, hjust = 1, vjust = 0.5),
        axis.text.y = element_text(size = 5),
    )

integration_scores_means_plot <- ggplot(
    integration_summary_plotting,
    aes(x = Type, y = Method, colour = Mean)
) +
    geom_point(shape = "square") +
    scale_colour_viridis_c(
        option = "magma",
        limits = c(0, max(integration_summary_plotting$Mean))
    ) +
    labs(
        title = "Scores",
        colour = "Mean score"
    ) +
    facet_wrap(~ Integration, nrow = 1) +
    theme_features_integration_extended +
    theme(
        axis.ticks.y = element_line(size = 0.25)
    )

integration_scores_diffs_means_plot <- ggplot(
    integration_diff_means_plotting,
    aes(x = Type, y = Method, colour = Mean)
) +
    geom_point(shape = "square") +
    colorspace::scale_colour_continuous_diverging(
        palette = "PurpleGreen",
    ) +
    facet_wrap(~ Integration, nrow = 1) +
    labs(
        title = "Difference in scores",
        colour = "Mean difference"
    ) +
    theme_features_integration_extended +
    theme(
        axis.text.y = element_blank(),
    )

integration_ranks_means_plot <- ggplot(
    integration_ranks_means_plotting,
    aes(x = Type, y = Method, colour = Type)
) +
    geom_point(
        aes(alpha = length(unique(Method)) -  MeanRank),
        shape = "square"
    ) +
    scale_y_discrete(drop = FALSE) +
    scale_colour_manual(values = types_palette, guide = "none") +
    scale_alpha_continuous(
        limits = c(length(unique(integration_ranks_means$Method)), 1),
        range = c(0.1, 1.0),
        trans = "reverse",
        breaks = scales::breaks_extended(n = 6)
    ) +
    facet_wrap(~ Integration, nrow = 1) +
    labs(
        title = "Ranks",
        alpha = "Mean rank"
    ) +
    guides(
        alpha = guide_legend(
            theme = theme(legend.text.position = "bottom"),
            nrow = 1
        )
    ) +
    theme_features_integration_extended +
    theme(
        axis.text.y = element_blank()
    )

integration_diffs_ranks_means_plot <- ggplot(
    integration_ranks_diff_means_plotting,
    aes(x = Type, y = Method, colour = Mean)
) +
    geom_point(shape = "square") +
    colorspace::scale_colour_continuous_diverging(
        palette = "Tropic",
        rev = TRUE
    ) +
    facet_wrap(~ Integration, nrow = 1) +
    labs(
        title = "Difference in ranks",
        colour = "Mean difference"
    ) +
    theme_features_integration_extended +
    theme(
        axis.text.y = element_blank(),
    )

integration_means_plot <- wrap_plots(
    integration_scores_means_plot,
    integration_scores_diffs_means_plot,
    integration_ranks_means_plot,
    integration_diffs_ranks_means_plot,
    nrow = 1,
    widths = c(3, 2, 3, 2)
)

integration_means_plot
```

## Integration SDs

```{r extended-integration-sds}
integration_scores_sds <- ggplot(
    integration_summary_plotting,
    aes(x = Type, y = Method, colour = SD)
) +
    geom_point(shape = "square") +
    scale_colour_viridis_c(option = "cividis") +
    labs(
        title = "Scores",
        colour = "Standard deviation"
    ) +
    facet_wrap(~ Integration, nrow = 1) +
    theme_features_integration_extended +
    theme(
        axis.ticks.y = element_line(size = 0.25)
    )

integration_scores_diffs_sds <- ggplot(
    integration_diff_means_plotting,
    aes(x = Type, y = Method, colour = SD)
) +
    geom_point(shape = "square") +
    scale_colour_viridis_c(option = "cividis") +
    facet_wrap(~ Integration, nrow = 1) +
    labs(
        title = "Difference in scores",
        colour = "Difference SD"
    ) +
    theme_features_integration_extended +
    theme(
        axis.text.y = element_blank(),
    )

integration_ranks_sds <- ggplot(
    integration_ranks_means_plotting,
    aes(x = Type, y = Method, colour = SDRank)
) +
    geom_point(shape = "square") +
    scale_y_discrete(drop = FALSE) +
    scale_colour_viridis_c(option = "cividis") +
    facet_wrap(~ Integration, nrow = 1) +
    labs(
        title = "Ranks",
        colour = "Rank SD"
    ) +
    theme_features_integration_extended +
    theme(
        axis.text.y = element_blank()
    )

integration_diffs_ranks_sds <- ggplot(
    integration_ranks_diff_means_plotting,
    aes(x = Type, y = Method, colour = SD)
) +
    geom_point(shape = "square") +
    scale_colour_viridis_c(option = "cividis") +
    facet_wrap(~ Integration, nrow = 1) +
    labs(
        title = "Difference in ranks",
        colour = "Difference SD"
    ) +
    guides(
        colour = guide_colourbar(order = 1),
    ) +
    theme_features_integration_extended +
    theme(
        axis.text.y = element_blank(),
    )

integration_sds_plot <- wrap_plots(
    integration_scores_sds,
    integration_scores_diffs_sds,
    integration_ranks_sds,
    integration_diffs_ranks_sds,
    nrow = 1,
    widths = c(3, 2, 3, 2)
)

integration_sds_plot
```

# Additional

Other figures

## Metric ranks

```{r supplementary-ranks, fig.height = 6, out.width = "100%"}
metric_ranks <- metrics_summary |>
    filter(Integration == "scVI-1") |>
    pivot_longer(
        cols = Classification:Overall,
        names_to = "Type",
        values_to = "Value"
    ) |>
    group_by(Dataset, Type) |>
    mutate(Rank = rank(Value)) |>
    mutate(
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        ),
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        )
    )
 
ranks_plot <- ggplot(metric_ranks) +
    ggforce::geom_sina(
        aes(x = length(unique(Method)) - Rank + 1, y = Method),
        size = 0.5, alpha = 0.3
    ) +
    geom_point(
        data = metric_ranks_means,
        aes(
            x = length(unique(Method)) - MeanRank + 1,
            y = Method,
            fill = length(unique(Method)) - MeanRank + 1
        ),
        shape = "circle filled", size = 2
    ) +
    scale_fill_viridis_c(direction = -1) +
    scale_x_reverse(
        limits = c(length(unique(metric_ranks$Method)) + 1, 1),
        breaks = c(1, 10, 20, 30)
    ) +
    facet_grid(. ~ Type, scales = "free_y", space = "free_y") +
    labs(
        title = "Method ranks by metric type",
        x = "Method rank",
        fill = "Mean rank"
    ) +
    theme_features_pub() +
    theme(
        axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        panel.spacing.y = unit(0, "cm"),
    )

ranks_plot
```

## Selected feature overlap

```{r supplementary-overlaps, fig.height = 12, out.width = "100%"}
overlaps_all_plotting <- overlaps |>
    mutate(
        Method1 = factor(
            Method1,
            levels = overlaps_clust_order,
            labels = str_remove(
                str_remove(
                    method_names[overlaps_clust_order_formatted],
                    "N=2000,?\\s?"
                ),
                "\\s\\(\\)"
            )
        ),
        Method2 = factor(
            Method2,
            levels = overlaps_clust_order,
            labels = str_remove(
                str_remove(
                    method_names[overlaps_clust_order_formatted],
                    "N=2000,?\\s?"
                ),
                "\\s\\(\\)"
            )
        ),
        Dataset = factor(
            Dataset,
            levels = names(dataset_names),
            labels = dataset_names
        )
    )

dataset_overlaps_plot <- ggplot(overlaps_all_plotting) +
    geom_tile(aes(x = Method2, y = Method1, fill = Jaccard)) +
    scale_fill_viridis_c(option = "plasma", limits = c(0, 1)) +
    scale_x_discrete(expand = c(0, 0)) +
    scale_y_discrete(expand = c(0, 0)) +
    coord_fixed() +
    facet_wrap(~ Dataset, nrow = 4) +
    labs(
        title = "Overlap in selected features by dataset",
        fill = "Jaccard\nIndex"
    ) +
    theme_features_pub() +
    theme(
        legend.position = "bottom",
        axis.title = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        panel.grid = element_blank()
    )

dataset_overlaps_plot
```

## Integration scores

```{r supplementary-integration-scores, fig.height = 9, out.width = "100%"}
integration_summary_long <- integration_summary |>
    pivot_longer(
        cols = scVI:Symphony,
        names_to = "Integration",
        values_to = "Value"
    ) |>
    mutate(
        Integration = factor(
            Integration,
            levels = c("scVI", "scANVI", "Symphony")
        ),
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        ),
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        )
    )

integration_scores_all_plot <- ggplot(integration_summary_long) +
    annotate(
        "rect",
        xmin = -Inf, xmax = 0, ymin = -Inf, ymax = Inf,
        fill = "red", alpha = 0.1
    ) +
    annotate(
        "rect",
        xmin = 1, xmax = Inf, ymin = -Inf, ymax = Inf,
        fill = "blue", alpha = 0.1
    ) +
    geom_vline(xintercept = 0, colour = "red") +
    geom_vline(xintercept = 1, colour = "blue") +
    ggforce::geom_sina(
        aes(x = Value, y = Method),
        size = 1, alpha = 0.3
    ) +
    geom_point(
        data = integration_summary_plotting,
        aes(x = Mean, y = Method, fill = Mean),
        shape = "circle filled", size = 1.5
    ) +
    scale_fill_viridis_c(option = "magma") +
    scale_x_continuous(limits = c(-3, 3), breaks = -3:3) +
    facet_grid(Integration ~ Type) +
    labs(
        title = "Metric summary scores by integration",
        x = "Metric type summary score",
        fill = "Mean score",
        caption = paste(
            "Note:", sum(abs(integration_summary_long$Value) > 3),
            "values outside the range [-3, 3] are not shown"
        )
    ) +
    theme_features_pub() +
    theme(
        axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
    )

integration_scores_all_plot
```

## Integration scores differences

```{r supplementary-integration-scores-diffs, fig.height = 6, out.width = "100%"}
integration_diffs_long <- integration_diffs |>
    select(-scVI, -scANVI, -Symphony) |>
    pivot_longer(
        cols = scANVIDiff:SymphonyDiff,
        names_to = "Integration",
        values_to = "Difference"
    ) |>
    mutate(
        Integration = str_remove(Integration, "Diff")
    ) |>
    mutate(
        Integration = factor(
            Integration,
            levels = c("scANVI", "Symphony")
        ),
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        ),
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        )
    )

integration_diffs_all_plot <- ggplot(integration_diffs_long) +
    geom_vline(xintercept = 0, colour = "red") +
    ggforce::geom_sina(
        aes(x = Difference, y = Method),
        size = 1, alpha = 0.3
    ) +
    geom_point(
        data = integration_diff_means_plotting,
        aes(x = Mean, y = Method, fill = Mean),
        shape = "circle filled", size = 1.5
    ) +
    colorspace::scale_fill_continuous_diverging(
        palette = "Purple-Green"
    ) +
    scale_x_continuous(limits = c(-3, 3), breaks = -3:3) +
    facet_grid(Integration ~ Type) +
    labs(
        title = "Difference in metric summary scores by integration",
        x = "Difference in metric type summary score",
        fill = "Mean difference",
        caption = paste(
            "Note:", sum(abs(integration_diffs_long$Difference) > 3),
            "values outside the range [-3, 3] are not shown"
        )
    ) +
    theme_features_pub() +
    theme(
        axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
    )

integration_diffs_all_plot
```

## Integration ranks

```{r supplementary-integration-ranks, fig.height = 9, out.width = "100%"}
integration_ranks_plotting <- integration_ranks |>
    mutate(
        Integration = factor(
            Integration,
            levels = c("scVI-1", "scANVI-1", "Symphony-1"),
            labels = c("scVI", "scANVI", "Symphony")
        ),
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        ),
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        )
    )

integration_ranks_all_plot <- ggplot(integration_ranks_plotting) +
    ggforce::geom_sina(
        aes(
            x = length(unique(integration_ranks$Method)) - Rank + 1,
            y = Method
        ),
        size = 1, alpha = 0.3
    ) +
    geom_point(
        data = integration_ranks_means_plotting,
        aes(
            x = length(unique(integration_ranks$Method)) - MeanRank + 1,
            y = Method,
            fill = MeanRank
        ),
        shape = "circle filled", size = 1.5
    ) +
    scale_fill_viridis_c() +
    scale_x_reverse(
        limits = c(length(unique(integration_ranks$Method)) + 1, 0),
        breaks = c(1, 10, 20, 30)
    ) +
    facet_grid(Integration ~ Type) +
    labs(
        title = "Metric summary score ranks by integration",
        x = "Metric type summary score rank",
        fill = "Mean rank",
    ) +
    theme_features_pub() +
    theme(
        axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
    )

integration_ranks_all_plot
```

## Integration ranks differences

```{r supplementary-integration-ranks-diffs, fig.height = 6, out.width = "100%"}
integration_rank_diffs_long <- integration_rank_diffs |>
    select(-scVI, -scANVI, -Symphony) |>
    pivot_longer(
        cols = scANVIDiff:SymphonyDiff,
        names_to = "Integration",
        values_to = "Difference"
    ) |>
    mutate(
        Integration = str_remove(Integration, "Diff")
    ) |>
    mutate(
        Integration = factor(
            Integration,
            levels = c("scANVI", "Symphony")
        ),
        Method = factor(
            Method,
            levels = methods_order,
            labels = method_names[methods_order]
        ),
        Type = factor(
            Type,
            levels = c("Overall", names(type_names)),
            labels = c("Overall", type_names)
        )
    )

integration_rank_diffs_all_plot <- ggplot(integration_rank_diffs_long) +
    geom_vline(xintercept = 0, colour = "red") +
    ggforce::geom_sina(
        aes(x = Difference, y = Method),
        size = 1, alpha = 0.3
    ) +
    geom_point(
        data = integration_ranks_diff_means_plotting,
        aes(x = Mean, y = Method, fill = Mean),
        shape = "circle filled", size = 1.5
    ) +
    colorspace::scale_fill_continuous_diverging(
        palette = "Tropic",
        rev = TRUE
    ) +
    facet_grid(Integration ~ Type) +
    labs(
        title = "Difference in metric summary score ranks by integration",
        x = "Difference in metric type summary score rank",
        fill = "Mean difference",
    ) +
    theme_features_pub() +
    theme(
        axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
    )

integration_rank_diffs_all_plot
```

# Output

Save output files

```{r output}
save_figure_files(
    benchmark_figure,
    here::here("analysis", "output", "benchmark"),
    width = 7, height = 7.5
)

save_figure_files(
    overlaps_panel,
    here::here("analysis", "output", "benchmark-overlaps"),
    width = 7, height = 3
)

save_figure_files(
    integration_figure,
    here::here("analysis", "output", "integration"),
    width = 7, height = 6
)

save_figure_files(
    category_scores_ranks_plot,
    here::here("analysis", "output", "category-scores-ranks"),
    width = 7, height = 5.5
)

save_figure_files(
    overlaps_means_sd_plot,
    here::here("analysis", "output", "overlaps-means-sds"),
    width = 7, height = 5
)

save_figure_files(
    dataset_overlaps_plot,
    here::here("analysis", "output", "dataset-overlaps"),
    height = 2.5
)

save_figure_files(
    integration_means_plot,
    here::here("analysis", "output", "integration-means"),
    width = 7, height = 4
)

save_figure_files(
    integration_sds_plot,
    here::here("analysis", "output", "integration-sds"),
    width = 7, height = 4
)

save_figure_files(
    ranks_plot,
    here::here("analysis", "output", "metric-ranks"),
    height = 6
)

save_figure_files(
    integration_scores_all_plot,
    here::here("analysis", "output", "integration-scores"),
    height = 9
)

save_figure_files(
    integration_diffs_all_plot,
    here::here("analysis", "output", "integration-score-diffs"),
    height = 6
)

save_figure_files(
    integration_ranks_all_plot,
    here::here("analysis", "output", "integration-ranks"),
    height = 9
)

save_figure_files(
    integration_rank_diffs_all_plot,
    here::here("analysis", "output", "integration-rank-diffs"),
    height = 6
)
```

# Session info {.unnumbered}

```{r session-info}
sessioninfo::session_info()
```
